asset-service/requirements.txt
# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
pymongo==4.6.1
motor==3.3.2  # Async MongoDB driver

# Object Storage
minio==7.2.3

# Authentication
PyJWT==2.8.0
python-jose[cryptography]==3.3.0

# File Processing
python-magic==0.4.27
Pillow==10.2.0
pydub==0.25.1
mutagen==1.47.0
ffmpeg-python==0.2.0

# Utilities
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0
aiofiles==23.2.1

# Logging
structlog==24.1.0

# HTTP Client
httpx==0.26.0
requests==2.31.0

# Date/Time
python-dateutil==2.8.2

# Validation
email-validator==2.1.0

# CORS
fastapi-cors==0.0.6

